Check for smoothness; not yet passing `smooth enough` test.
