+++ floris/tools/turbine_utilities.py
@@ -176,3 +176,17 @@ def _find_nearest_value_for_wind_speed(test_vals, ws_vals, ws):
     errs = np.absolute(ws_vals-ws)
     idx = errs.argmin()
     return test_vals[idx]
+
+def check_smooth_power_curve(power):
+    """
+    Check whether there are "wiggles" in the power signal.
+    """
+
+    tol = 0.01 # kW, tolerance per point
+
+    dir_changes = np.sum(np.abs(np.diff(np.sign(np.diff(power)))))
+    is_smooth = dir_changes <= 2
+
+    import ipdb; ipdb.set_trace()
+
+    return is_smooth
